Fixed login adding new file loginFunction.js
