+++ src/loginFunctions.js
@@ -1,0 +1,108 @@
+import fire from "./fire.js";
+import Hero from "./hero.js";
+import React, { useState, useEffect } from "react";
+import Login from"./login";
+
+
+const LoginFunctions = () => {
+
+const [user, setUser] = useState("");
+const [email, setEmail] = useState("");
+const [password, setPassword] = useState("");
+const [emailError, setEmailError] = useState("");
+const [passwordError, setPasswordError] = useState("");
+const [hasAccount, setHasAccount] = useState(false);
+
+const clearInput = () => {
+  setEmail("");
+  setPassword("");
+};
+
+const clearErrors = () => {
+  setEmailError("");
+  setPasswordError("Please enter at least 6 characters");
+};
+
+const handleLogin = () => {
+  clearErrors();
+  fire
+    .auth()
+    .signInWithEmailAndPassword(email, password)
+    .catch((err) => {
+      switch (err.code) {
+        case "auth/email-already-in-use":
+        case "auth/invalid-email":
+          setEmailError(err.message);
+          break;
+        case "auth/weak-password":
+          setPasswordError(err.message);
+          break;
+      }
+    });
+};
+
+const handleSignup = () => {
+  clearErrors();
+  fire
+    .auth()
+    .createUserWithEmailAndPassword(email, password)
+    .catch((err) => {
+      switch (err.code) {
+        case "auth/invalid-email":
+        case "auth/user-disabled":
+          setEmailError(err.message);
+          break;
+        case "auth/wrong-password":
+          setPasswordError(err.message);
+          break;
+      }
+    });
+
+};
+
+const handleLogout = () => {
+  fire.auth().signOut();
+};
+
+const authListener = () => {
+  fire.auth().onAuthStateChanged((user) => {
+    if (user) {
+      clearInput();
+      setUser(user);
+    } else {
+      setUser("");
+    }
+  });
+};
+
+useEffect(() => {
+  authListener();
+}, []);
+
+    return (
+
+        <div className="Login">
+        {user ? (
+          <Hero handleLogout={handleLogout} />
+        ) : (
+          <Login
+            email={email}
+            setEmail={setEmail}
+            password={password}
+            setPassword={setPassword}
+            handleLogin={handleLogin}
+            handleSignup={handleSignup}
+            hasAccount={hasAccount}
+            setHasAccount={setHasAccount}
+            emailError={emailError}
+            passwordError={passwordError}
+          />
+        )}
+      </div>
+    )
+
+    
+
+}
+
+export default LoginFunctions;
